+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 - Added a requeue button to the party dequeued message and /rfurequeue command.
 - Prevent sending party/chat messages when muted and added /rfuunmuteme command.
+- Added Moby-Duck Wisdom buff HUD display.
 
 ### Fixes
 - Fixed date input in /rfudrophistory add not working.